@@ -3,29 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@daily-co/daily-js": "^0.9.981",
-    "@stripe/react-stripe-js": "^1.1.0",
-    "@stripe/stripe-js": "^1.2.0",
-    "antd": "^3.7.3",
+    "@daily-co/daily-js": "^0.9.992",
+    "@stripe/react-stripe-js": "^1.1.2",
+    "@stripe/stripe-js": "^1.8.0",
+    "antd": "^3.26.18",
     "axios": "^0.18.0",
     "babel-core": "^6.26.3",
-    "babel-loader": "^8.0.6",
+    "babel-loader": "^8.1.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "draft-js": "^0.11.6",
     "express": "^4.17.1",
-    "moment": "^2.24.0",
+    "moment": "^2.27.0",
     "node-sass": "^4.14.1",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-draft-wysiwyg": "^1.14.5",
     "react-facebook-login": "^4.1.1",
     "react-google-login": "^5.1.21",
-    "react-infinite-scroll-component": "^5.0.4",
+    "react-infinite-scroll-component": "^5.0.5",
     "react-redux": "^5.0.7",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "^3.3.1",
-    "react-select-country-list": "^2.1.2",
+    "react-scripts": "^3.4.1",
+    "react-select-country-list": "^2.2.1",
     "redux": "^4.0.0",
     "redux-thunk": "^2.3.0"
   },
@@ -47,22 +47,22 @@
     "npm": "6.11.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.7.4",
-    "@babel/plugin-proposal-class-properties": "^7.7.4",
-    "@babel/preset-env": "^7.7.4",
-    "@babel/preset-react": "^7.7.4",
+    "@babel/core": "^7.11.1",
+    "@babel/plugin-proposal-class-properties": "^7.10.4",
+    "@babel/preset-env": "^7.11.0",
+    "@babel/preset-react": "^7.10.4",
     "babel-plugin-import": "^1.13.0",
     "clean-webpack-plugin": "^3.0.0",
-    "css-loader": "^3.2.0",
-    "file-loader": "^5.0.2",
+    "css-loader": "^3.6.0",
+    "file-loader": "^5.1.0",
     "html-webpack-plugin": "^3.2.0",
-    "less": "^3.10.3",
+    "less": "^3.12.2",
     "less-loader": "^5.0.0",
     "less-vars-to-js": "^1.3.0",
     "mini-css-extract-plugin": "^0.9.0",
-    "react-app-rewired": "^2.1.5",
-    "style-loader": "^1.0.1",
-    "terser-webpack-plugin": "^2.3.4",
+    "react-app-rewired": "^2.1.6",
+    "style-loader": "^1.2.1",
+    "terser-webpack-plugin": "^2.3.7",
     "webpack-merge": "^4.2.2"
   },
   "browserslist": {
