@@ -1,6 +1,7 @@
 from rest_framework import serializers
 from articlesApi.models import Article, Video, Comment, Category, Like, Rating, Image, CommentReply, FeedbackTypes
-from users.models import User
+from users.models import User, ProfileInfo
+from users.serializers import ProfileSerializer, ProfileInfoSerializer
 from django.core.files.storage import FileSystemStorage
 import json
 
@@ -18,17 +19,30 @@ class CommentReplySerializer(serializers.ModelSerializer):
 class ArticleSerializer(serializers.ModelSerializer):
     engagement = StringSerializer(many=True)
     categories = StringSerializer(many=True)
+    overview = StringSerializer(many=False)
     author = StringSerializer(many=False)
     video = StringSerializer(many=False)
+    user_thumbnail = serializers.SerializerMethodField()
+    user_name = serializers.SerializerMethodField()
     
+    def get_user_thumbnail(self, obj):
+        request = self.context.get('request')
+        print("COLAS: ", obj.author.profile.profile_avatar)
+        profile_thumbnail = ProfileSerializer(obj.author.profile, many=False, context={'request': request}).data["profile_avatar"]
+        print("COLAS 2: ", profile_thumbnail)
+        return profile_thumbnail
+
+    def get_user_name(self, obj):
+        profile_name = ProfileInfoSerializer(ProfileInfo.objects.get(profile_username=obj.author.id), many=False ).data.get("name")
+        return profile_name
 
     class Meta:
         model = Article
-        fields = ('id', 'title', 'content', 'timestamp',
+        fields = ('id', 'title', 'overview', 'content', 'timestamp',
                   'engagement', 'categories', 'author',
-                  "comment_count", "view_count", "rating_count",
+                  "comment_count", "rating_count",
                     "likes_count", "view_count", "rating_count", "avg_rating",
-                    "video", "thumbnail"
+                    "video", "thumbnail", "user_thumbnail", "user_name"
                   )
 
     def get_feedback_forms(self, obj):
