@@ -74,20 +74,22 @@ class EditorContainer extends Component{
     //     <h3 className="loading">Loading...</h3>
     //   );
     // }
-    return <div className='editor'>
-      <Editor
-        editorState={editorState}
-        onEditorStateChange={this.onEditorStateChange.bind(this)}    
-        toolbar={{
-          inline: { inDropdown: true },
-          list: { inDropdown: true },
-          textAlign: { inDropdown: true },
-          link: { inDropdown: true },
-          history: { inDropdown: true },
-          // image: { uploadCallback: uploadImageCallBack, alt: { present: true, mandatory: true } },
-        }}
-      />
-    </div>
+    return (
+          <div className='editor'>
+            <Editor
+              editorState={editorState}
+              onEditorStateChange={this.onEditorStateChange.bind(this)}    
+              toolbar={{
+                inline: { inDropdown: true },
+                list: { inDropdown: true },
+                textAlign: { inDropdown: true },
+                link: { inDropdown: true },
+                history: { inDropdown: true },
+                // image: { uploadCallback: uploadImageCallBack, alt: { present: true, mandatory: true } },
+              }}
+            />
+          </div>
+    )
   }
 }
 
